+++ front/src/api/author/hooks/get-author-list.tsx
@@ -1,26 +1,24 @@
 import { axios } from "@/lib/axios";
 import { useInfiniteQuery } from "@tanstack/react-query";
-import { AuthorListDTO, AuthorList } from "./types";
+import { AuthorListDTO, AuthorList, QueryFnType } from "../types";
 import { InfiniteQueryConfig } from "@/lib/react-query";
+import { authorQuery } from "..";
 
 const REQUEST_SIZE = 10;
 
-async function getAuthorList(params: AuthorListDTO): Promise<AuthorList> {
-  return axios.get(`/user/list`, { params });
-}
-
-type QueryFnType = typeof getAuthorList;
 type UseAuthorListOptions = {
   params: AuthorListDTO;
-  config?: InfiniteQueryConfig<QueryFnType>;
+  config?: InfiniteQueryConfig<QueryFnType<AuthorListDTO>>;
 };
 
 export function useAuthorList({ params, config }: UseAuthorListOptions) {
+  const { queryKey, queryFn } = authorQuery.authorList();
+
   return useInfiniteQuery({
     ...config,
-    queryKey: ["AuthorList", params],
+    queryKey: queryKey(params),
     queryFn: ({ pageParam }) =>
-      getAuthorList({ ...params, size: REQUEST_SIZE, page: pageParam }),
+      queryFn({ ...params, size: REQUEST_SIZE, page: pageParam }),
     initialPageParam: 0,
     getNextPageParam: (lastPage, allPages) => {
       const morePagesExist = lastPage?.users?.length === REQUEST_SIZE;
