+++ front/src/api/author/index.ts
@@ -1,1 +1,2 @@
-export { useAuthorList } from "./get-author-list";
+export { useAuthorList } from "./hooks/get-author-list";
+export { authorQuery } from "./apis/author-api";
