+++ front/src/api/author/types.ts
@@ -4,6 +4,12 @@ export type AuthorListDTO = {
   size?: number;
 };
 
+export type FollowAuthorDTO = {
+  userId: number;
+  page?: number;
+  size?: number;
+};
+
 export type Author = {
   userId: number;
   userName: string;
@@ -16,3 +22,5 @@ export type Author = {
 export type AuthorList = {
   users: Author[];
 };
+
+export type QueryFnType<T> = (params: T) => Promise<AuthorList>;
